+++ src/wallet/hd_public.cpp
@@ -231,7 +231,7 @@ hd_public hd_public::derive_public(uint32_t index) const
     const hd_lineage lineage
     {
         lineage_.prefixes,
-        lineage_.depth + 1u,
+        static_cast<uint8_t>(lineage_.depth + 1),
         fingerprint(),
         index
     };
